fix compositColumn
If the columns parameter of a compositeColumn includes the $ character, it is automatically removed, and a warning log is generated to inform the developer that the specified fields must be actual fields belonging to the same table. For this reason, it is neither necessary nor appropriate to use the $ prefix on field names.