new hierarchy for logger, introducing the branch "gnr.audit" which does not propagate to ancestors, and provide a different interface for logging, exposing dynamic methods based on contextual statements, like:
from gnr.sql import sqlauditlogger

sqlauditlogger.delete(delete_query)

sqlauditlogger.insert(insert_query)

All Auditor loggers expects extra argument, including users, which is
defaulted to the local user (useful for CLI commands)

Added to the execute() the automatically insertion of the current genropy user
using the add_sql_comment decorator logic. Since this decorator was planned
to audit queries from the database log, which is a strange architecture choices, it can
be replaced by something similar called @audit.

Added to test handler for the auditor (just change the formatting output in a StreamHandler
to show the extra data as the user) and a netport handler example, which opens a listening
tcp socket where all the logs are sent.