change by @fporcari and @gporcari - tweaked db adapters loading process in order to provide a custom "sub-adapter" implementation for the connection.
Example configuration:

<db implementation="gnrpkg.neon.adapters.postgres" dbbranch="development"
    user="neondb_owner" dbname="application1" />

The database connection for the instance will be handled as a postgres connection
by the adapter defined in python module 'neon.adapters.postgres'.

If the newly user custom adapter needs any configuration, it's up to the
deployer/user to update it accordingly.

Added a new parameter to the database connection definition "dbbranch", to
specificy, when supported, the database branch to be used.