Execute the GLBL data check and load on onDbUpgrade from the 'nazione' table, since the simple upgrade script is not enough, since it's executed *after* the sys record creation, and the problematic relation is coming from a sys record.
The upgrade script is left there, just in case, to be used via `gnr db migrate -U`

We also had to tweak the adm.preference table and handle a case when the preference is missing:
if we call the loadStartupData for a package, a record from adm.preferences is update, but since
we're creating the database from scratch, it doesn't exist anymore. So we just skip the update
if the record in adm.preference is missing. YMMV, please review