refs #308
ugly refactoring to merge the old and new async task scheduler/worker
and being able to choose which implementation to be used, filling up the code
with condition to check which implementation has to be used.

using <tasks async_impl='true'/> in environment.xml will allow to start the new implementation.

PLEASE NOTE: all the services must restart after the environment change, since the old scheduler
is executed in the daemon, the workers are separate processes, and the siterunner uses table, views and
menu related to the new implementation.